Gibbs: Pass number of capabilities in explicitly
While this might be the right thing to do anyways, the motivation for
this was strange behavior from setNumCapabilities which seems to impose
a >10 second latency before new capabilities begin execution. By passing
in the number of capabilities explicitly, we can create needed
capabilities up front and only use a subset of them in the benchmark.